Refactor: Convert useToast hook to utility function
- useToast 훅을 showToast 유틸리티 함수로 변경
- Toast의 border 색상을 연하게 변경
