--- apps/web/src/_shared/helpers/utils/.gitkeep
@@ -1,1 +1,0 @@
-shared/helpers/utils 안에는 여러 페이지에서 재사용 가능한 일반 함수가 위치합니다.
