+++ apps/web/src/app/login/_helpers/hooks/useGoogleAuthFlow.ts
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 
-import { useToast } from '@/_shared/helpers/hooks/useToast';
+import { showToast } from '@/_shared/helpers/utils/showToast';
 import { useAuthorizationFlow } from '@/app/login/_helpers/hooks/useAuthorizationFlow';
 import { signInWithGoogle } from '@workspace/firebase/auth';
 
@@ -10,7 +10,6 @@ import { signInWithGoogle } from '@workspace/firebase/auth';
  */
 export const useGoogleAuthFlow = () => {
   const [isLoading, setIsLoading] = useState(false);
-  const { showToast } = useToast();
   const { mutateAsync: authorizationMutation } = useAuthorizationFlow();
 
   const loginWithGoogle = async () => {
@@ -30,7 +29,7 @@ export const useGoogleAuthFlow = () => {
       console.error('🚨 Google 로그인 실패 또는 토큰 처리 중 오류:', err);
 
       showToast({
-        message: 'Google 로그인에 실패했어요. 다시 시도해주세요.',
+        message: 'Google 로그인에 실패했어요',
         type: 'error',
       });
     } finally {
