+++ apps/web/src/app/register-club/success/_clientBoundary/QrCardClient/index.tsx
@@ -3,8 +3,8 @@
 import { useRef, useState } from 'react';
 
 import { APP_PATH } from '@/_shared/helpers/constants/appPath';
-import { useToast } from '@/_shared/helpers/hooks/useToast';
 import { buildUrlWithParams } from '@/_shared/helpers/utils/buildUrlWithParams';
+import { showToast } from '@/_shared/helpers/utils/showToast';
 import { RegisterClubSuccessCardFooter } from '@/app/register-club/success/_components/RegisterClubSuccessCardFooter';
 import { CardContent } from '@workspace/ui/components/card';
 import { Input } from '@workspace/ui/components/input';
@@ -23,13 +23,13 @@ export const QrCardClient = () => {
   const qrCodeRef = useRef<HTMLDivElement>(null);
 
   const router = useRouter();
-  const { showToast } = useToast();
 
   if (!clubEnglishName) {
     showToast({
       message: '유효하지 않은 동아리입니다.',
       type: 'error',
     });
+    
     router.replace(APP_PATH.REGISTER_CLUB.HOME);
 
     return null;
