+++ apps/web/src/app/sign-up/_helpers/hooks/useSignUpFlow.ts
@@ -1,6 +1,6 @@
 import { useForm } from 'react-hook-form';
 
-import { useToast } from '@/_shared/helpers/hooks/useToast';
+import { showToast } from '@/_shared/helpers/utils/showToast';
 import { zodResolver } from '@hookform/resolvers/zod';
 
 import { type UserProfileFormData } from '../types';
@@ -16,7 +16,6 @@ type UserProfile = {
 
 export const useSignUpFlow = () => {
   const { mutateAsync: registerProfileMutation } = useRegisterProfile();
-  const { showToast } = useToast();
 
   const form = useForm<UserProfileFormData>({
     resolver: zodResolver(userProfileSchema),
@@ -43,7 +42,7 @@ export const useSignUpFlow = () => {
       console.error('🚨 프로필 제출 중 오류 발생:', error);
 
       showToast({
-        message: '프로필 완성 중에 오류가 발생했어요 🥲 다시 시도해주세요',
+        message: '프로필 완성 중에 오류가 발생했어요',
         type: 'error',
       });
     }
